site: real validate and new Independence page, tone/coherence check

+++ src/models/losses.py
@@ -10,5 +10,9 @@ def joint_loss(model, x, target_train_space, corr_lbl, sepA_lbl, sepB_lbl, w_cls
     import torch.nn.functional as F
     mean, L, cl, al, bl = model.forward_all(x)
     nll = -torch.distributions.MultivariateNormal(mean, scale_tril=L).log_prob(target_train_space).mean()
-    ce = (F.cross_entropy(cl, corr_lbl) + F.cross_entropy(al, sepA_lbl) + F.cross_entropy(bl, sepB_lbl))
+    ce = (
+        F.cross_entropy(cl, corr_lbl.long()) + 
+        F.cross_entropy(al, sepA_lbl.long()) + 
+        F.cross_entropy(bl, sepB_lbl.long())
+    )
     return nll + w_cls * ce, nll.item(), ce.item()
